+++ src/main/java/tw/sales/taxes/tax/TaxCalculate.java
@@ -1,14 +1,18 @@
 package tw.sales.taxes.tax;
 
+import java.math.BigDecimal;
+
 import tw.sales.taxes.goods.Goods;
 
 /**
  * Tax calculate interface
+ * You should implement it if you want to add a new tax calculate.
+ * 
  * @author liuyang
- *
+ * 
  */
 public interface TaxCalculate {
-	
-	float calculateTax(Goods goods);
-	
+
+	BigDecimal calculateTax(Goods goods);
+
 }
