+++ src/main/java/tw/sales/taxes/tax/TaxCaluclate4Import.java
@@ -1,11 +1,19 @@
 package tw.sales.taxes.tax;
 
+import java.math.BigDecimal;
+
 import tw.sales.taxes.goods.Goods;
 
+/**
+ * Implements interface TaxCalculate
+ * 
+ * @author liuyang
+ * 
+ */
 public class TaxCaluclate4Import implements TaxCalculate {
 
 	@Override
-	public float calculateTax(Goods goods) {
+	public BigDecimal calculateTax(Goods goods) {
 		CompositeTax comTax = new CompositeTax();
 		comTax.addTax(new SingleTax(goods.getGoodsType().getBasicTaxRate()));
 		if (goods.isImported())
